testing different webworker configurations
